requirements.txt
# Phygrid Edge App Dependencies
phystack-hub-client>=0.2.0
